stores: remove conflict handling on applyUnappliedOutputAdditions
